Allow user to provide different http/https proxies
It is very possible that proxies will be different for http and https.
In this case, we would want the user to have the option to specify a
unique value for each.

This adds these knobs without removing the previous API, so as not to
break workflows that use overrides['proxy']

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>